refactor: extract shared markdown renderer to utils/markdown-renderer.ts
- Move customMarked (markedHighlight + hljs) and renderer (code/listitem/list) to shared module
- MemoCard and TimelinePost now import from /utils/markdown-renderer
- Deduplicate ~100 identical lines (was copy-pasted in both components)
- Behavior unchanged: same renderer, gfm is marked default
- Verified: svelte-check 0, eslint 0, tests 61/61
